+++ tests/test_casereader.py
@@ -5,11 +5,8 @@ import shutil
 import pytest
 
 from ansys.fluent.core import examples
-from ansys.fluent.core.filereader.casereader import (
-    CaseReader,
-    InputParameter,
-    _get_processed_string,
-)
+from ansys.fluent.core.filereader.case import _get_processed_string
+from ansys.fluent.core.filereader.casereader import CaseReader, InputParameter
 
 
 def call_casereader(
